+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -156,7 +156,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-04T09:24:01.774Z",
+    "2026-06-04T11:04:28.064Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {